feat: трекер задач может расставлять задачи по приоритету и проверять, не пересекаются ли они по времени выполнения

+++ src/service/InMemoryTaskManager.java
@@ -1,30 +1,28 @@
 package service;
 
+import exception.ValidateException;
 import model.Epic;
 import model.Subtask;
 import model.Task;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 public class InMemoryTaskManager implements TaskManager {
     protected final HistoryManager historyManager;
-    protected final Map<Integer, Task> tasks;
-    protected final Map<Integer, Epic> epics;
-    protected final Map<Integer, Subtask> subtasks;
+    protected final Map<Integer, Task> tasks = new HashMap<>();
+    protected final Map<Integer, Epic> epics = new HashMap<>();
+    protected final Map<Integer, Subtask> subtasks = new HashMap<>();
+
+    protected final Set<Task> prioritizedTasks = new TreeSet<>(Comparator.comparing(Task::getStartTime));
 
     protected int seq = 0;
 
     public InMemoryTaskManager(HistoryManager historyManager) {
         this.historyManager = historyManager;
-        this.tasks = new HashMap<>();
-        this.epics = new HashMap<>();
-        this.subtasks = new HashMap<>();
     }
 
 //    Методы для каждого из типа задач(Задача/Эпик/Подзадача):
-//    a. Получение списка всех задач.
+//    Получение списка всех задач.
     @Override
     public Collection<Task> getAllTasks() {
         return this.tasks.values();
@@ -40,10 +38,11 @@ public class InMemoryTaskManager implements TaskManager {
         return this.subtasks.values();
     }
 
-//    b. Удаление всех задач.
+//    Удаление всех задач.
     @Override
     public void removeAllTasks() {
         historyManager.removeAll(tasks.values());
+        prioritizedTasks.removeAll(tasks.values());
         tasks.clear();
     }
 
@@ -51,23 +50,22 @@ public class InMemoryTaskManager implements TaskManager {
     public void removeAllEpics() {
         historyManager.removeAll(epics.values());
         historyManager.removeAll(subtasks.values());
+        prioritizedTasks.removeAll(subtasks.values());
         epics.clear();
-        subtasks.clear(); // Каждая подзадача закреплена под эпиком
+        subtasks.clear();
     }
 
     @Override
     public void removeAllSubtasks() {
         historyManager.removeAll(subtasks.values());
-        // Сперва удалим все подзадачи в эпиках
-        for (Map.Entry<Integer, Epic> epicEntry : epics.entrySet()) {
-            Epic epic = epicEntry.getValue();
-            epic.deleteAllSubtasks();
-            epic.updateStatus();
-        }
+        prioritizedTasks.removeAll(subtasks.values());
+        epics.values().stream()
+                .peek(Epic::deleteAllSubtasks)
+                .forEach(Epic::update);
         subtasks.clear();
     }
 
-//    c. Получение по идентификатору.
+//    Получение по идентификатору.
     @Override
     public Task getTask(Integer id) {
         Task task = tasks.get(id);
@@ -89,11 +87,15 @@ public class InMemoryTaskManager implements TaskManager {
         return subtask;
     }
 
-//    d. Создание. Сам объект должен передаваться в качестве параметра.
+//    Создание. Сам объект должен передаваться в качестве параметра.
     @Override
     public Task createTask(Task task) {
         task.setId(generateId());
         tasks.put(task.getId(), task);
+        if (task.getStartTime() != null) {
+            validateTask(task);
+            prioritizedTasks.add(task);
+        }
         return task;
     }
 
@@ -113,8 +115,12 @@ public class InMemoryTaskManager implements TaskManager {
         }
         subtask.setId(generateId());
         epicFromManager.addSubtask(subtask);
-        epicFromManager.updateStatus();
+        epicFromManager.update();
         subtask.setEpic(epicFromManager);
+        if (subtask.getStartTime() != null) {
+            validateTask(subtask);
+            prioritizedTasks.add(subtask);
+        }
         subtasks.put(subtask.getId(), subtask);
         return subtask;
     }
@@ -123,13 +129,15 @@ public class InMemoryTaskManager implements TaskManager {
         return ++seq;
     }
 
-//    e. Обновление. Новая версия объекта с верным идентификатором передаётся в виде параметра.
+//    Обновление. Новая версия объекта с верным идентификатором передаётся в виде параметра.
     @Override
     public void updateTask(Task task) {
         Task saved = tasks.get(task.getId());
         if (saved == null) {
             return;
         }
+        updateTaskTime(saved, task);
+
         saved.setName(task.getName());
         saved.setDescription(task.getDescription());
         saved.setStatus(task.getStatus());
@@ -151,35 +159,52 @@ public class InMemoryTaskManager implements TaskManager {
         if (saved == null) {
             return;
         }
+        updateTaskTime(saved, subtask);
+
         saved.setName(subtask.getName());
         saved.setDescription(subtask.getDescription());
         saved.setStatus(subtask.getStatus());
         if (!saved.getEpic().equals(subtask.getEpic())) {
             Epic oldEpic = saved.getEpic();
             oldEpic.deleteSubtask(saved);
-            oldEpic.updateStatus();
+            oldEpic.update();
 
             Epic newEpic = subtask.getEpic();
             newEpic.addSubtask(saved);
             saved.setEpic(newEpic);
         }
 
-        saved.getEpic().updateStatus();
+        saved.getEpic().update();
     }
 
-//    f. Удаление по идентификатору.
+    private void updateTaskTime(Task saved, Task newTask) {
+        if (newTask.getStartTime() != null && newTask.getDuration() != null) {
+            validateTask(newTask);
+            prioritizedTasks.remove(saved);
+            saved.setStartTime(newTask.getStartTime());
+            saved.setDuration(newTask.getDuration());
+            prioritizedTasks.add(saved);
+        } else if (saved.getStartTime() != null) { // Если время начала было удалено
+            prioritizedTasks.remove(saved);
+        }
+    }
+
+//    Удаление по идентификатору.
     @Override
     public void deleteTask(int id) {
         historyManager.remove(id);
-        tasks.remove(id);
+        Task deleted = tasks.remove(id);
+        if (deleted != null && deleted.getStartTime() != null) {
+            prioritizedTasks.remove(deleted);
+        }
     }
 
     @Override
     public void deleteEpic(int id) {
         historyManager.remove(id);
-        Collection<Subtask> epicSubtasks = getEpicSubtasks(id);
-        for (Subtask subtask : epicSubtasks) {
+        for (Subtask subtask : getEpicSubtasks(id)) {
             historyManager.remove(subtask.getId());
+            prioritizedTasks.remove(subtask);
             subtasks.remove(subtask.getId());
         }
         epics.remove(id);
@@ -188,15 +213,19 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public void deleteSubtask(int id) {
         historyManager.remove(id);
-        Subtask deleted = subtasks.get(id);
-        Epic epic = deleted.getEpic();
-        epic.deleteSubtask(deleted);
-        epic.updateStatus();
-        subtasks.remove(id);
+        Subtask deleted = subtasks.remove(id);
+        if (deleted != null) {
+            if (deleted.getStartTime() != null) {
+                prioritizedTasks.remove(deleted);
+            }
+            Epic epic = deleted.getEpic();
+            epic.deleteSubtask(deleted);
+            epic.update();
+        }
     }
 
 //    Дополнительные методы:
-//    a. Получение списка всех подзадач определённого эпика.
+//    Получение списка всех подзадач определённого эпика.
     @Override
     public Collection<Subtask> getEpicSubtasks(int epicId) {
         Collection<Subtask> subtasks = epics.get(epicId).getSubtasks();
@@ -207,6 +236,22 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public Collection<Task> getHistory() {
         return historyManager.getAll();
+    }
+
+    protected void validateTask(Task task) {
+        Optional<Task> inter = prioritizedTasks.stream().filter(temp -> !Objects.equals(temp.getId(), task.getId()) &&
+                !temp.getStartTime().isAfter(task.getEndTime()) &&
+                !task.getStartTime().isAfter(temp.getEndTime())
+        ).findAny();
+
+        if (inter.isPresent()) {
+            throw new ValidateException("Пересечение с задачей: " + inter.get());
+        }
+    }
+
+    @Override
+    public Collection<Task> getPrioritizedTasks() {
+        return prioritizedTasks;
     }
 
 }
